Add support for anime streaming with vlc
